Move contact link in navbar to its own group
Previously, the contact link was hidden under "Association". From feedback from members it has become apparent that this is rather confusing and leads to people searching the entire menu for that contact link. To make it easier to access that (rather crucial) link, I'm putting it back to the top level of the menu.